Extract common parts from order note mutations
